move killable_intents support to base class
