+++ ovos_workshop/skills/ovos.py
@@ -192,111 +192,6 @@ class OVOSSkill(MycroftSkill):
         time.sleep(0.5)  # if TTS had not yet started
         self.bus.emit(msg.forward("mycroft.audio.speech.stop"))
 
-    # abort get_response gracefully
-    def _wait_response(self, is_cancel, validator, on_fail, num_retries):
-        """Loop until a valid response is received from the user or the retry
-        limit is reached.
-
-        Arguments:
-            is_cancel (callable): function checking cancel criteria
-            validator (callbale): function checking for a valid response
-            on_fail (callable): function handling retries
-
-        """
-        self._response = False
-        self._real_wait_response(is_cancel, validator, on_fail, num_retries)
-        while self._response is False:
-            time.sleep(0.1)
-        return self._response
-
-    def __get_response(self):
-        """Helper to get a reponse from the user
-
-        Returns:
-            str: user's response or None on a timeout
-        """
-
-        def converse(utterances, lang=None):
-            converse.response = utterances[0] if utterances else None
-            converse.finished = True
-            return True
-
-        # install a temporary conversation handler
-        self.make_active()
-        converse.finished = False
-        converse.response = None
-        self.converse = converse
-
-        # 10 for listener, 5 for SST, then timeout
-        # NOTE a threading event is not used otherwise we can't raise the
-        # AbortEvent exception to kill the thread
-        start = time.time()
-        while time.time() - start <= 15 and not converse.finished:
-            time.sleep(0.1)
-            if self._response is not False:
-                if self._response is None:
-                    # aborted externally (if None)
-                    self.log.debug("get_response aborted")
-                converse.finished = True
-                converse.response = self._response  # external override
-        self.converse = self._original_converse
-        return converse.response
-
-    def _handle_killed_wait_response(self):
-        self._response = None
-        self.converse = self._original_converse
-
-    @killable_event("mycroft.skills.abort_question", exc=AbortQuestion,
-                    callback=_handle_killed_wait_response, react_to_stop=True)
-    def _real_wait_response(self, is_cancel, validator, on_fail, num_retries):
-        """Loop until a valid response is received from the user or the retry
-        limit is reached.
-
-        Arguments:
-            is_cancel (callable): function checking cancel criteria
-            validator (callbale): function checking for a valid response
-            on_fail (callable): function handling retries
-
-        """
-        num_fails = 0
-        while True:
-            if self._response is not False:
-                # usually None when aborted externally
-                # also allows overriding returned result from other events
-                return self._response
-
-            response = self.__get_response()
-
-            if response is None:
-                # if nothing said, prompt one more time
-                num_none_fails = 1 if num_retries < 0 else num_retries
-                if num_fails >= num_none_fails:
-                    self._response = None
-                    return
-            else:
-                # catch user saying 'cancel'
-                if is_cancel(response):
-                    self._response = None
-                    return
-                validated = validator(response)
-                # returns the validated value or the response
-                # (backwards compat)
-                if validated is not False and validated is not None:
-                    self._response = response if validated is True else validated
-                    return
-
-            num_fails += 1
-            if 0 < num_retries < num_fails or self._response is not False:
-                self._response = None
-                return
-
-            line = on_fail(response)
-            if line:
-                self.speak(line, expect_response=True)
-            else:
-                self.bus.emit(Message('mycroft.mic.listen',
-                                      context={"skill_id": self.skill_id}))
-
 
 # backwards compat alias, no functional difference
 class OVOSFallbackSkill(OVOSSkill):
